fixed the closing in gallery
