+++ frontend/src/components/home/style.css
@@ -217,3 +217,15 @@
     margin-top: 20px;
   }
 }
+
+.modal-container {
+  position: fixed;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background: rgba(0, 0, 0, 0.8);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
